Complete refactor of entities.
- Completely refactored how entities work.
- Removed redundant classes like IEntity.
- Added BaseEntityMovement and BaseEntity.
- Fixed/Removed all missing imports and redundant code.
