+++ Assets/Scripts/Entities/Abilities/GlideBehaviour.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using Interfaces;
-using UnityEngine;
+﻿using UnityEngine;
 
 public class GlideBehaviour : MonoBehaviour
 {
