+++ Assets/Scripts/Entities/Animals/CatBehaviour.cs
@@ -1,18 +1,17 @@
-﻿using Enums;
 using System;
 using System.Collections.Generic;
 using Entities;
+using Enums;
 using UnityEngine;
 using UnityEngine.AI;
 
-public class VillagerBehaviour : BaseEntity, IPossessable
+public class CatBehaviour : BaseEntity
 {
-    void Awake()
+    private void Awake()
     {
-        // Todo give Name and Profession
         Rigidbody = GetComponent<Rigidbody>();
         NavMeshAgent = GetComponent<NavMeshAgent>();
-
+        
         FearThreshold = 20;
         FearDamage = 0;
         FaintDuration = 10;
@@ -20,17 +19,19 @@ public class VillagerBehaviour : BaseEntity, IPossessable
         ScaredOfGameObjects = new Dictionary<Type, float>()
         {
             [typeof(RatBehaviour)] = 3f,
-            [typeof(ILevitateable)] = 3f
+            [typeof(PoliceManBehaviour)] = 2f,
+            [typeof(VillagerBehaviour)] = 2f,
+            [typeof(ILevitateable)] = 5f
         };
     }
 
-    void Update()
+    private void Update()
     {
         CheckSurroundings();
     }
-
+    
     public override void UseFirstAbility()
     {
-        //TODO Villager first ability.
+        //TODO first ability.
     }
 }
