+++ Assets/Scripts/Entities/BaseEntity.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 26feedc2e54440098779067200dd10ec
+timeCreated: 1605868052
