+++ Assets/Scripts/Entities/BaseEntityMovement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5fc0f27446fc49b7b59183f08c31481d
+timeCreated: 1605868042
