+++ Assets/Scripts/Entities/BaseMovement.cs
@@ -6,39 +6,16 @@ using UnityEngine.AI;
 
 public abstract class BaseMovement : MonoBehaviour
 {
-    public GameObject Target;
-    public NavMeshAgent NavMeshAgent;
     public Rigidbody Rigidbody;
-
-    public float MinimumFollowRange, MaximumFollowRange;
     public float Speed;
     public bool IsGrounded = true;
-    public bool IsOnCountdown;
+    public float JumpForce = 10.0f;
+    public Collider Collider;
 
-    [SerializeField] private PathFindingState _pathFindingState;
     private float _rotationSpeed = 10f;
     private float _gravity = 9.81f;
-    public float JumpForce = 10.0f;
     private bool _hasCollidedWithWall;
-    public Collider Collider;
     private ContactPoint[] _contacts;
-    private bool _isPathFinding;
-    private bool _hasPositionInArea;
-    private Quaternion _spawnRotation;
-    private Vector3 _spawnLocation;
-    private Vector3 _patrolDestination;
-    private EntityArea _currentArea;
-
-    private void Start()
-    {
-        if (NavMeshAgent)
-        {
-            NavMeshAgent.autoBraking = true;
-
-            _spawnRotation = transform.rotation;
-            _spawnLocation = transform.position;
-        }
-    }
 
     public void MoveEntityInDirection(Vector3 direction, float speed)
     {
@@ -88,27 +65,11 @@ public abstract class BaseMovement : MonoBehaviour
         }
     }
 
-    public void MoveEntityInDirection(Vector3 direction)
+    public virtual void MoveEntityInDirection(Vector3 direction)
     {
         MoveEntityInDirection(direction, Speed);
     }
 
-    public void MoveWithPathFinding()
-    {
-        switch (_pathFindingState)
-        {
-            case PathFindingState.Stationary:
-                ReturnToSpawn();
-                break;
-            case PathFindingState.Patrolling:
-                PatrolArea();
-                break;
-            case PathFindingState.Following:
-                FollowTarget();
-                break;
-        }
-    }
-    
     public void Jump()
     {
         IsGrounded = false;
@@ -155,77 +116,5 @@ public abstract class BaseMovement : MonoBehaviour
         }
 
         IsGrounded = false;
-    }
-
-    private void FollowTarget()
-    {
-        if (Target)
-        {
-            float distanceToTarget = Vector3.Distance(transform.position, Target.transform.position);
-            if (distanceToTarget > MinimumFollowRange && distanceToTarget < MaximumFollowRange)
-            {
-                NavMeshAgent.isStopped = false;
-                NavMeshAgent.SetDestination(Target.transform.position);
-                return;
-            }
-
-            NavMeshAgent.isStopped = true;
-        }
-    }
-
-    private void ReturnToSpawn()
-    {
-        if (HasReachedDestination(_spawnLocation))
-        {
-            NavMeshAgent.isStopped = false;
-            NavMeshAgent.destination = _spawnLocation;
-            return;
-        }
-        
-        Quaternion lerpToRotation = Quaternion.Lerp(transform.rotation, _spawnRotation, 
-            Time.deltaTime * 5f);
-        transform.rotation = lerpToRotation;
-    }
-
-    private void PatrolArea()
-    {
-        if (!_currentArea) MoveToNextArea();
-
-        if (!_hasPositionInArea)
-        {
-            _patrolDestination = EntityAreaHandler.Instance.GetRandomPositionInArea(_currentArea, gameObject);
-            NavMeshAgent.destination = _patrolDestination;
-            _hasPositionInArea = true;
-        }
-
-        if (HasReachedDestination(_patrolDestination))
-        {
-            _hasPositionInArea = false;
-        }
-    }
-
-    public void ChangePathFindingState(PathFindingState pathFindingState)
-    {
-        _pathFindingState = pathFindingState;
-    }
-
-    private bool HasReachedDestination(Vector3 destination)
-    {
-        float distanceToDestination = Vector3.Distance(transform.position, destination);
-        return distanceToDestination < 0.5f;
-    }
-
-    public IEnumerator StartCountdownInArea(float amountOfTime)
-    {
-        yield return new WaitForSeconds(amountOfTime);
-        _currentArea = null;
-        IsOnCountdown = false;
-    }
-
-    private void MoveToNextArea()
-    {
-        _currentArea = EntityAreaHandler.Instance.GetRandomAreaForEntity(gameObject);
-        _hasPositionInArea = false;
-        NavMeshAgent.ResetPath();
     }
 }
