--- Assets/Scripts/Entities/Humans/Police.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 00129e178a23e7e45b40dd46b7a3f668
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
