--- Assets/Scripts/Interfaces/Entities.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a1105e116e1249e9874583ad7d679d6b
-timeCreated: 1603879031
