--- Assets/Scripts/Interfaces/Entities/IAnimal.cs
@@ -1,7 +1,0 @@
-namespace Interfaces
-{
-    public interface IAnimal : IEntity
-    {
-        //Animal distinct things.
-    }
-}
