--- Assets/Scripts/Interfaces/Entities/IAnimal.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a45800c7b3eb4b399f726fde19462c6b
-timeCreated: 1603879002
