--- Assets/Scripts/Interfaces/Entities/IEntity.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 9e0bd9cf749949709d3a58a0b4ccc4e1
-timeCreated: 1603878813
