--- Assets/Scripts/Interfaces/Entities/IHuman.cs
@@ -1,7 +1,0 @@
-namespace Interfaces
-{
-    public interface IHuman : IEntity
-    {
-        //Human distinct things.
-    }
-}
