--- Assets/Scripts/Interfaces/Entities/IHuman.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f3d2a651b24f424aae28b4ec56b007e2
-timeCreated: 1603878973
