--- Assets/Scripts/Interfaces/IEmotion.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 7ca573d5978f46968b0407a8f64388ca
-timeCreated: 1603878850
