+++ Assets/Scripts/Sandbox/PossessableSanboxBox.cs
@@ -1,68 +1,19 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using Interfaces;
-using Enums;
+﻿using Entities;
 
-public class PossessableSanboxBox : BaseMovement, IEntity, IPossessable
+public class PossessableSanboxBox : BaseEntity, IPossessable
 {
     private CameraController _cameraController;
-
-    public float FearThreshold { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-    public float FearDamage { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-    public float FaintDuration { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-    public EmotionalState EmotionalState { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-    public Dictionary<Type, float> ScaredOfGameObjects { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-    public bool IsPossessed { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-    public Dialogue Dialogue => throw new NotImplementedException();
-
-    public Question Question => throw new NotImplementedException();
-
-    public CharacterList CharacterName { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-    public List<CharacterList> Relationships => throw new NotImplementedException();
-
-    public Sentence[] DefaultAnswers { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
     private void Awake()
     {
         _cameraController = UnityEngine.Camera.main.GetComponent<CameraController>();
     }
 
-    private void Update()
+    public override void UseFirstAbility()
     {
-    }
-
-    public IEnumerator CalmDown()
-    {
-        yield return null;
-    }
-
-    public void EntityJump()
-    {
-        throw new NotImplementedException();
-    }
-
-    public void CheckSurroundings()
-    {
-    }
-
-    public void DealFearDamage(float amount)
-    {
-    }
-
-    public void Faint()
-    {
-    }
-
-    public void Move(Vector3 direction)
-    {
-        MoveEntityInDirection(direction);
-    }
-
-    public void UseFirstAbility()
-    {
+        /* TODO PossessableSanboxBox ability. I mean... tf... wtf is a PossessableSanboxBox...
+         * I mean wtf it isn't even spelled correctly. PossessableSANBOXBox. Seriously?
+         * Who pays for this shit?
+         */
     }
 }
